requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy
opencv-python
Pillow==6.2.1
pymongo==3.9.0
six==1.13.0
soupsieve==1.9.5
Werkzeug==0.16.0
selenium
pillow
requests